separating the order details components
