+++ ccSchwabManager/Views/HoldingsView/PositionDetailView/PositionDetailContent/OrderTab/CurrentOrdersSection/OrderGroupHeightKey.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+struct OrderGroupHeightKey: PreferenceKey {
+    static var defaultValue: CGFloat = 0
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value = max(value, nextValue())
+    }
+}
